pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "attractor"
version = "0.1.0"
description = "DOT-based AI pipeline runner with unified LLM client and coding agent loop"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=24.0.0",
    "jsonschema>=4.23.0",
]

[project.scripts]
attractor = "cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true
